@@ -21,16 +21,18 @@ from src.backend.functions import retrieve_custom_from_url
 from src.backend.main import create_classes_from_custom_id
 
 
-def get_last_tasks():
+def get_last_tasks(module):
     custom_id = retrieve_custom_from_url(request)
     _vars = create_classes_from_custom_id(custom_id)
     database = _vars[0]
+
     tasks = database.select({
         'select': ['*', "to_char(creation_date, 'HH24:MI:SS') as begin_time",
                    "to_char(end_date, 'HH24:MI:SS') as end_time",
                    '(Extract(epoch FROM (CURRENT_TIMESTAMP - creation_date))/60)::INTEGER as age'],
         'table': ['tasks_watcher'],
-        'where': ["status IS NULL OR creation_date >= NOW() - INTERVAL '1 hour'"],
+        'where': ['module = %s', "(status IS NULL OR creation_date >= NOW() - INTERVAL '1 hour')"],
+        'data': [module],
         'order_by': ["id desc"],
     })
 
