@@ -16,13 +16,12 @@
 # @dev : Oussama Brich <oussaba.brich@edissyum.com>
 
 from src.backend.import_controllers import tasks_watcher
-from flask import Blueprint, make_response, jsonify
-
+from flask import Blueprint, make_response, jsonify, request
 
 bp = Blueprint('task_watcher', __name__, url_prefix='/ws/')
 
 
 @bp.route('tasks/progress', methods=['GET'])
 def get_last_task():
-    tasks = tasks_watcher.get_last_tasks()
+    tasks = tasks_watcher.get_last_tasks(request.args['module'])
     return make_response(jsonify(tasks[0])), tasks[1]
